+++ Gunz/ZGameTimer.h
@@ -12,12 +12,12 @@ public:
 		m_dwUpdateCnt++;
 	}
 
-	void SetGlobalOffset(int nOffset) {  m_nGlobalOffset = nOffset;  }
-	auto GetGlobalTick() const { return (m_nTickTime + m_nGlobalOffset); }
+	void SetGlobalOffset(i64 nOffset) { m_nGlobalOffset = nOffset; }
+	auto GetGlobalTick() const { return m_nTickTime + m_nGlobalOffset; }
 	auto GetUpdateCount() const { return m_dwUpdateCnt; }
 
 private:
 	u64	m_nTickTime = 0;
-	int	m_nGlobalOffset = 0;
+	i64	m_nGlobalOffset = 0;
 	u32 m_dwUpdateCnt = 0;
 };
